+++ src-tauri/src/utils/mod.rs
@@ -2,8 +2,10 @@ mod format;
 mod progress;
 mod extract;
 mod path;
+mod command;
 
 pub use format::*;
 pub use progress::*;
 pub use extract::*;
 pub use path::*;
+pub use command::*;